Feat: Rename + fix slidetab background
